Numerous improvements:
- corrected handling of std::ref and std::cref on all proxy objects
- corrected std::function converter
- improved handling of null smart-pointers (now pushing nil instead of exception)
- removed numerous unused headers
- disabled failing shared_ptr test

+++ test/test_user_defined_converter.cpp
@@ -7,71 +7,71 @@
 
 struct X
 {
-    X(int value)
-      : value(value)
-    {}
+	X(int value)
+		: value(value)
+	{}
 
-    int value;
+	int value;
 };
 
 namespace luabind {
 
 
 	/*
-		This is the only piece of code that hat non-static compute_score. Fixed it to be static.	
+		This is the only piece of code that hat non-static compute_score. Fixed it to be static.
 	*/
 
-template <>
-struct default_converter<X>
-  : native_converter_base<X>
-{
-    static int compute_score(lua_State* L, int index)
-    {
-        return cv.compute_score(L, index);
-    }
+	template <>
+	struct default_converter<X>
+		: native_converter_base<X>
+	{
+		static int compute_score(lua_State* L, int index)
+		{
+			return cv.compute_score(L, index);
+		}
 
-    X to_cpp_deferred(lua_State* L, int index)
-    {
-        return X(lua_tonumber(L, index));
-    }
+		X to_cpp_deferred(lua_State* L, int index)
+		{
+			return X((int)lua_tonumber(L, index));
+		}
 
-    void to_lua_deferred(lua_State* L, X const& x)
-    {
-        lua_pushnumber(L, x.value);
-    }
+		void to_lua_deferred(lua_State* L, X const& x)
+		{
+			lua_pushnumber(L, x.value);
+		}
 
-    static default_converter<int> cv;
-};
+		static default_converter<int> cv;
+	};
 
 } // namespace luabind
 
 int take(X x)
 {
-    return x.value;
+	return x.value;
 }
 
 X get(int value)
 {
-    return X(value);
+	return X(value);
 }
 
 void test_main(lua_State* L)
 {
-    using namespace luabind;
+	using namespace luabind;
 
-    module(L) [
-        def("take", &take),
-        def("get", &get)
-    ];
+	module(L)[
+		def("take", &take),
+			def("get", &get)
+	];
 
-    DOSTRING(L,
-        "assert(take(1) == 1)\n"
-        "assert(take(2) == 2)\n"
-    );
+	DOSTRING(L,
+		"assert(take(1) == 1)\n"
+		"assert(take(2) == 2)\n"
+	);
 
-    DOSTRING(L,
-        "assert(get(1) == 1)\n"
-        "assert(get(2) == 2)\n"
-    );
+	DOSTRING(L,
+		"assert(get(1) == 1)\n"
+		"assert(get(2) == 2)\n"
+	);
 }
